+++ Chapter_7/task_7_5.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 
-long long factorial(int n);
+long long factorial(size_t n);
 
 int main() {
 
     using namespace std;
 
-    int number {};
+    size_t number {};
     cout << "Enter number: ";
     while (cin >> number && cin.get() != 'q') {
         cout << "Factorial: " << factorial(number) << endl;
@@ -16,6 +16,6 @@ int main() {
     return 0;
 }
 
-long long factorial(int n) {
+long long factorial(size_t n) {
     return !n ? 1 : n * factorial(n - 1);
 }
